# server/internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for managing assignments, questions, classrooms and more.",
        "title": "ParableMind API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/subjectives": {
            "get": {
                "tags": [
                    "Subjectives"
                ],
                "summary": "Get all Subjectives",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bank ID",
                        "name": "bankId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/questions.Subjective"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subjectives"
                ],
                "summary": "Create Subjective",
                "parameters": [
                    {
                        "description": "Subjective Question",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/questions.Subjective"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/questions.Subjective"
                        }
                    }
                }
            }
        },
        "/subjectives/bulk": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subjectives"
                ],
                "summary": "Bulk Create Subjectives",
                "parameters": [
                    {
                        "description": "List of Subjective Questions",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/questions.Subjective"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/questions.Subjective"
                            }
                        }
                    }
                }
            }
        },
        "/subjectives/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subjectives"
                ],
                "summary": "Get Subjective by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Subjective ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/questions.Subjective"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subjectives"
                ],
                "summary": "Update Subjective",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Subjective",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/questions.Subjective"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/questions.Subjective"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Subjectives"
                ],
                "summary": "Delete Subjective",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Subjective ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subjectives"
                ],
                "summary": "Patch Subjective",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updates",
                        "name": "updates",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Variable": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "namePositions": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "value": {
                    "type": "string"
                },
                "valuePositions": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "variableType": {
                    "type": "string"
                }
            }
        },
...
        "questions.Subjective": {
            "type": "object",
            "properties": {
                "bankId": {
                    "type": "string"
                },
                "gradingCriteria": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "idealAnswer": {
                    "type": "string"
                },
                "points": {
                    "type": "integer"
                },
                "question": {
                    "type": "string"
                },
                "variable": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Variable"
                    }
                }
            }
        }
    }
}